+++ RELEASE-NOTES.md
@@ -5,6 +5,7 @@
 * Switching MrPlow to continue using the same channel
 * Switching from maven-jar to new maven-install based system for bazel 2.0
 * Added watch-only note on wallet summary page of iceleaf GUI
+* Add details button to iceleaf GUI to get xpubs and seeds of wallets
 
 ## 1.6.0